Table Component in Current Account Page
