+++ content-scripts/pages/class_page.tsx
@@ -1,6 +1,7 @@
 import { fetchSigarraPage } from "../modules/utilities/pageUtils";
 import { getPath } from "../modules/utilities/sigarra";
 import { Table } from "../components/Table";
+import { extractTableData } from "../modules/table";
 
 const removeExtras = () => {
     const heading = document.querySelector("#conteudoinner > h2");
@@ -62,56 +63,6 @@ const createPhotosButton = (icon: Element, url: string, classIndex: number) => {
 const getPhotosLink = (title: HTMLElement): string => {
     const linkElement = title.children[2] as HTMLAnchorElement;
     return linkElement.href;
-};
-
-// Helper function to extract table data from HTML table
-const extractTableData = (
-    table: HTMLElement,
-): { headers: [string, string | Element][]; data: (string | Element)[][] } => {
-    const headers: [string, string | Element][] = [];
-    const data: (string | Element)[][] = [];
-
-    const headerRow =
-        table.querySelector("thead tr") || table.querySelector("tr");
-    if (headerRow) {
-        const headerCells = headerRow.querySelectorAll("th, td");
-        headerCells.forEach((cell, index) => {
-            const key = `col_${index}`;
-            const value = cell.textContent?.trim() || "";
-            headers.push([key, value]);
-        });
-    }
-
-    const tbody = table.querySelector("tbody") || table;
-    const rows = tbody.querySelectorAll("tr");
-
-    const startIndex = table.querySelector("thead") ? 0 : 1;
-
-    for (let i = startIndex; i < rows.length; i++) {
-        const row = rows[i];
-        const cells = row.querySelectorAll("td, th");
-        const rowData: (string | Element)[] = [];
-
-        cells.forEach((cell) => {
-            if (cell.children.length > 0) {
-                if (cell.children.length === 1) {
-                    rowData.push(cell.firstElementChild!);
-                } else {
-                    const wrapper = document.createElement("span");
-                    wrapper.innerHTML = cell.innerHTML;
-                    rowData.push(wrapper);
-                }
-            } else {
-                rowData.push(cell.textContent?.trim() || "");
-            }
-        });
-
-        if (rowData.length > 0) {
-            data.push(rowData);
-        }
-    }
-
-    return { headers, data };
 };
 
 const editTitle = async (
